refactor: LottoNumber의 캐시 관리를 자신이 하도록 변경
