+++ README.md
@@ -30,6 +30,8 @@
 # 뷰 기능 목록
 - [X] 구입금액을 입력받는 기능
     - [X] [예외1] 입력값이 정수인지 확인하는 기능
+- [ ] 수동으로 구입할 로또 개수를 입력받는 기능
+    - [ ] [예외1] 입력값이 정수인지 확인하는 기능
 - [X] 몇 개의 로또티켓을 구매했는지 출력하는 기능
 - [X] 구입한 티켓의 번호들을 출력하는 기능
 - [X] 당첨 및 보너스 번호를 입력하는 기능
