+++ src/main/java/lotto/controller/LottoController.java
@@ -55,7 +55,7 @@ public class LottoController {
 
 	private static LottoNumber createBonusNumber() {
 		try {
-			return AllLottoNumbers.getLottoNumber(InputView.inputBonusNumber());
+			return LottoNumber.getLottoNumber(InputView.inputBonusNumber());
 		} catch (LottoTicketIllegalArgumentException e) {
 			OutputView.printWarningMessage(e.getMessage());
 			return createBonusNumber();
