+++ src/test/java/lotto/domain/MockLottoNumberGenerator.java
@@ -5,6 +5,6 @@ import java.util.List;
 public class MockLottoNumberGenerator implements RandomGenerator {
 	@Override
 	public List<LottoNumber> generateSixNumbers() {
-		return AllLottoNumbers.getAll().subList(0, 6);
+		return LottoNumber.getAll().subList(0, 6);
 	}
 }
