+++ src/test/java/lotto/domain/RandomLottoTicketFactoryTest.java
@@ -19,7 +19,7 @@ public class RandomLottoTicketFactoryTest {
 
 		// then
 		List<LottoNumber> lottoNumbers = Stream.of(1, 2, 3, 4, 5, 6)
-				.map(AllLottoNumbers::getLottoNumber)
+				.map(LottoNumber::getLottoNumber)
 				.collect(Collectors.toList());
 		Assertions.assertThat(randomLottoTicket).isEqualTo(new SerialLottoNumber(lottoNumbers));
 	}
